moved to separate ConvertSprintfCallsPass pipe
