+++ compiler/compiler.cmake
@@ -123,6 +123,7 @@ prepend(KPHP_COMPILER_PIPES_SOURCES pipes/
         collect-main-edges.cpp
         collect-required-and-classes.cpp
         convert-list-assignments.cpp
+        convert-sprintf-calls.cpp
         erase-defines-declarations.cpp
         extract-async.cpp
         extract-resumable-calls.cpp
